+++ src/queue.c
@@ -57,7 +57,7 @@ void generateQueueDetails() {
     for(uint32_t queueIndex = 1; queueIndex < queueCount; queueIndex++) {
         VkBool32 queueDistinct = VK_TRUE;
 
-        for(int32_t comparisonIndex = queueIndex - 1; comparisonIndex >= 0; comparisonIndex--) {
+        for(int32_t comparisonIndex = (int32_t) queueIndex - 1; comparisonIndex >= 0; comparisonIndex--) {
             if(queueReferences[queueIndex]->queueFamilyIndex == queueReferences[comparisonIndex]->queueFamilyIndex) {
                 queueReferences[queueIndex]->queueInfoIndex = queueReferences[comparisonIndex]->queueInfoIndex;
                 queueReferences[queueIndex]->queueIndex = queueReferences[comparisonIndex]->queueIndex + 1;
